CRITICAL FIX: All warnings fixed + URL input bug resolved
Fixed ALL build warnings:
- src/ai/cllm_training.c: Fixed sign comparison warnings (cast to size_t)
- app/crawler_thread.c: Fixed format truncation (limited string length)
- app/main.c: Fixed format truncation (limited string length)
- app/training_thread.c: Fixed format truncation (limited string length)
- app/workspace.c: Added missing header include
- app/ui/tabs/tab_training.c: Marked unused function with __attribute__((unused))

CRITICAL URL INPUT FIX:
- Root cause: Button handler was reading from state->crawler_start_url which was empty
- State was only updated when input was deactivated, but button click happened before deactivation
- Solution: Button handler now reads URL DIRECTLY from crawler_url_input.text
- This ensures the URL is always available regardless of input activation state
- Also updates state after reading for future use

The crawler will now use the URL you paste, not Wikipedia or empty string.

+++ src/ai/cllm_training.c
@@ -205,7 +205,7 @@ int cllm_load_training_data(CLLMTraining* training, const char* filename) {
     if (!training->model->tokens) {
         fprintf(stderr, "Warning: Model has no vocabulary, using character-based tokenization\n");
         // Fallback: character-based tokenization
-        for (size_t i = 0; i < bytes_read && tokens_added < file_size; i++) {
+        for (size_t i = 0; i < bytes_read && tokens_added < (size_t)file_size; i++) {
             if (content[i] != '\n' && content[i] != '\r') {
                 training->tokens[old_num_tokens + tokens_added] = (uint32_t)(content[i] % training->model->vocab_size);
                 tokens_added++;
@@ -214,7 +214,7 @@ int cllm_load_training_data(CLLMTraining* training, const char* filename) {
     } else {
         // Use vocabulary-based tokenization
         char* token = strtok(content, " \n\t");
-        while (token != NULL && tokens_added < file_size) {
+        while (token != NULL && tokens_added < (size_t)file_size) {
             // Find token in vocabulary
             bool found = false;
             for (uint32_t i = 0; i < training->model->vocab_size; i++) {
